fix(effect): keep Effect-typed programs off the public declaration graph
